refactor: :recycle: AppModule 의 Root Module 들을 분리, database root module 별도 폴더로 분리
테스트 환경에서도 RootModule 이 필요한 경우가 많을 것으로 예상되어 db module 분리하는 김에 같이 했습니다.

- close #312
