+++ app/src/api/coalition/db/coalition.database.aggregate.ts
@@ -1,7 +1,7 @@
 import {
   CollectionLookup,
   lookupStage,
-} from 'src/common/db/common.db.aggregation';
+} from 'src/database/mongoose/database.mongoose.aggregation';
 // eslint-disable-next-line
 import type { coalition } from './coalition.database.schema';
