+++ app/src/api/exam/exam.service.ts
@@ -4,7 +4,7 @@ import type { Model } from 'mongoose';
 import {
   findOneAndLean,
   type QueryOneArgs,
-} from 'src/common/db/common.db.query';
+} from 'src/database/mongoose/database.mongoose.query';
 import { exam } from './db/exam.database.schema';
 
 export const EXAM_PROJECT_IDS = [1320, 1321, 1322, 1323, 1324];
