+++ app/src/api/projectSession/db/projectSession.database.aggregate.ts
@@ -1,7 +1,7 @@
 import {
   CollectionLookup,
   lookupStage,
-} from 'src/common/db/common.db.aggregation';
+} from 'src/database/mongoose/database.mongoose.aggregation';
 
 /**
  *
