+++ app/src/api/score/db/score.database.aggregate.ts
@@ -2,7 +2,7 @@ import type { FilterQuery } from 'mongoose';
 import {
   CollectionLookup,
   lookupStage,
-} from 'src/common/db/common.db.aggregation';
+} from 'src/database/mongoose/database.mongoose.aggregation';
 import type { UserRank } from 'src/common/models/common.user.model';
 import type { DateRange } from 'src/dateRange/dtos/dateRange.dto';
 import type { score } from './score.database.schema';
