+++ app/src/config/database.ts
@@ -6,17 +6,24 @@ export type DatabaseConfig = {
   PASSWORD: string;
   ENDPOINT: string;
   NAME: string;
+  CONNECTION_STRING: string;
 };
 
 export const DATABASE_CONFIG = 'database';
+
+const USERNAME = findEnvByKey('DB_USERNAME');
+const PASSWORD = findEnvByKey('DB_PASSWORD');
+const ENDPOINT = findEnvByKey('DB_ENDPOINT');
+const NAME = findEnvByKey('DB_NAME');
 
 export const databaseConfig = registerAs(
   DATABASE_CONFIG,
   () =>
     ({
-      USERNAME: findEnvByKey('DB_USERNAME'),
-      PASSWORD: findEnvByKey('DB_PASSWORD'),
-      ENDPOINT: findEnvByKey('DB_ENDPOINT'),
-      NAME: findEnvByKey('DB_NAME'),
+      USERNAME,
+      PASSWORD,
+      ENDPOINT,
+      NAME,
+      CONNECTION_STRING: `mongodb://${USERNAME}:${PASSWORD}@${ENDPOINT}/${NAME}`,
     } satisfies DatabaseConfig),
 );
